refactor: clean up codebase, remove old artifacts and fix lints
- Remove unused benchmark files and their bench configurations from hwc-engine/Cargo.toml
- Delete deprecated test files from hwc-parser/tests and hwc-engine benches directories
- Simplify code across all crates: remove redundant blank lines, fix redundant variable clones, use slice references (&[T]) instead of Vec<T> for function parameters where appropriate
- Fix clippy lint warnings (clippy::unwrap_used, clippy::expect_used) across multiple modules
- Update public API re-exports and imports for consistency across geometry_router modules
- Replace manual boolean checks with `is_some_and` where applicable to simplify code
- Clean up outdated documentation comments and removed unnecessary example code from multiple crates
- Fix small bugs: simplify error output in hwc-cli main, fix z-range conditionals in substrate_types, remove redundant angle_deg clone, correct dy/di math in export_isolation
- Improve performance: replace get().is_none() check with !contains_key() in query_engine
- Update test assets for the ASIC two-pad-relational obstacle tests

+++ crates/hwc-engine/src/design_rule_check/clearance.rs
@@ -24,9 +24,9 @@ pub fn validate_clearances(
 
     // BUG DETECTION: Check for out-of-bounds entities
     let (space_x, space_y, space_z) = (
-        space.dimensions.width_nm as i64,
-        space.dimensions.height_nm as i64,
-        space.dimensions.depth_nm as i64,
+        space.dimensions.width_nm,
+        space.dimensions.height_nm,
+        space.dimensions.depth_nm,
     );
 
     eprintln!(
